[ issue #105 ] NCDG@K first draft + NDCG@10 refactoring
